+++ spec/components/back_link_spec.rb
@@ -13,11 +13,16 @@ describe "Back Link", type: :view do
 
   it "renders a back link correctly" do
     render_component(href: '/back-me')
-    assert_select ".govuk-back-link[href=\"/back-me\"]", text: "Back"
+    assert_select '.govuk-back-link[href="/back-me"]', text: "Back"
   end
 
   it "can render in welsh" do
     I18n.with_locale(:cy) { render_component(href: '/back-me') }
-    assert_select ".govuk-back-link[href=\"/back-me\"]", text: "Yn ôl"
+    assert_select '.govuk-back-link[href="/back-me"]', text: "Yn ôl"
+  end
+
+  it "renders a back link with data attributes" do
+    render_component(href: '/back-me', data_attributes: { tracking: "GT-123" })
+    assert_select '.govuk-back-link[data-tracking="GT-123"]'
   end
 end